Move task_configurator related classess into separate files
