+++ cpp/gpu/include/epseon/gpu/libgpu.hpp
@@ -4,12 +4,18 @@
 
 #include "epseon/gpu/predecl.hpp"
 
-#include "epseon/gpu/algorithms/algorithm.hpp"
-#include "epseon/gpu/algorithms/vibwa.hpp"
 #include "epseon/gpu/common.hpp"
 #include "epseon/gpu/compute_context.hpp"
 #include "epseon/gpu/device_interface.hpp"
 #include "epseon/gpu/enums.hpp"
-#include "epseon/gpu/python/api.hpp"
-#include "epseon/gpu/task_configurator.hpp"
 #include "epseon/gpu/task_handle.hpp"
+
+#include "epseon/gpu/python/api.hpp"
+
+#include "epseon/gpu/algorithms/algorithm.hpp"
+#include "epseon/gpu/algorithms/vibwa.hpp"
+
+#include "epseon/gpu/task_configurator/algorithm_config.hpp"
+#include "epseon/gpu/task_configurator/hardware_config.hpp"
+#include "epseon/gpu/task_configurator/potential_source.hpp"
+#include "epseon/gpu/task_configurator/task_configurator.hpp"
