+++ cpp/gpu/source/epseon/gpu/python/api.cpp
@@ -3,7 +3,7 @@
 #include "epseon/gpu/common.hpp"
 #include "epseon/gpu/compute_context.hpp"
 #include "epseon/gpu/enums.hpp"
-#include "epseon/gpu/task_configurator.hpp"
+#include "epseon/gpu/task_configurator/task_configurator.hpp"
 #include "fmt/format.h"
 #include "pybind11/detail/common.h"
 #include "pybind11/pytypes.h"
